Move BERTTextTransform into MUGEN example
[ghstack-poisoned]
